finalize keycloak implementation, see commit for details
- make sure to remove Keycloak session when logging out

- make sure rudimentary auth approach works

- add a default user in the development realm config (still annoying that we have to go through the UI and create the realm...)

- reorganize the application in a way that the ugly auth implementation conditionals are confined to __init__.py files and that specific implementations never actually share a file
