+++ .dockerignore
@@ -6,6 +6,6 @@
 !uv.lock
 !alembic.ini
 !migrations
-!realm-export.json
 # but make sure to ignore cache
 __pycache__
+*.pyc
