+++ .env.docker-example
@@ -2,6 +2,7 @@
 
 LOG_LEVEL=DEBUG
 PRODUCTION=true
+DOMAIN="http://localhost:8000"
 BASE_URL=
 
 # Application
@@ -9,6 +10,9 @@ BASE_URL=
 #DEVELOPMENT_API_KEY=blahblahblah
 SECRET_NAME=changemechangeme
 AUTH_IMPLEMENTATION=rudimentary
+
+# Keycloak Auth specific
+# TODO - we need to add valid Docker configurations here
 
 # INTERSECT
 # these values should be publicized
