+++ .env.example
@@ -2,23 +2,21 @@
 # Note that a .env file is not necessary if all of these values are environment variables
 
 LOG_LEVEL=DEBUG
+DOMAIN="http://localhost:8000"
 BASE_URL=
 
 # Application
 # You will generally not need to uncomment the DEVELOPMENT_API_KEY
 #DEVELOPMENT_API_KEY=blahblahblah
 SECRET_NAME=changemechangeme
+# change this to "keycloak" to test out keycloak
 AUTH_IMPLEMENTATION=rudimentary
 
 # Keycloak Auth specific
-AUTHORIZE_URL="http://127.0.0.1:8080/realms/DevRegistryKeycloak/protocol/openid-connect/auth"
-TOKEN_URL="http://127.0.0.1:8080/realms/DevRegistryKeycloak/protocol/openid-connect/token"
-JWKS_URL="http://127.0.0.1:8080/realms/DevRegistryKeycloak/protocol/openid-connect/certs"
+KEYCLOAK_REALM_BASE_URL="http://127.0.0.1:8080/realms/DevRegistryKeycloak/protocol/openid-connect"
 SCOPE="openid email"
 CLIENT_ID="registry-service-dev"
-CLIENT_SECRET="secret"
-REDIRECT_URL="http://localhost:8000/api/v1/auth/callback"
-
+CLIENT_SECRET="oOwbABbnGEwRlXU3HzCgspBtWfwIiIfQ"
 SESSION_SECRET="base64string (32/64 chars)"
 SESSION_FINGERPRINT_COOKIE="registry__Secure_Fgp"
 SESSION_VERIFY_ID=False
