--- intersect_registry_service/app/auth/impl/__init__.py
@@ -1,23 +1,0 @@
-from collections.abc import Callable
-
-from ...core.environment import settings
-from ..session import session_manager
-from ..user import USER
-
-get_user: Callable[[str], USER | None] = None  # type: ignore[assignment]
-"""
-This function should return any credentials necessary.
-
-For the rudimentary approach, it's returning the username and the password, though we probably don't want the password.
-"""
-
-if settings.AUTH_IMPLEMENTATION == 'keycloak':
-    from .keycloak import get_user_impl
-
-    get_user = get_user_impl
-else:
-    from .rudimentary import get_user_impl
-
-    get_user = get_user_impl
-
-session_manager.user_loader()(get_user)
