--- intersect_registry_service/app/auth/user.py
@@ -1,6 +1,0 @@
-USER = tuple[str, str]
-"""
-The defined typing is the return value of what the LoginManager looksup.
-
-For now it's just a simple username/password tuple, may want to change this later (especially once we start defining roles).
-"""
