+++ intersect_registry_service/app/ui/templates/pages/keycloak-login-page.jinja
@@ -1,0 +1,11 @@
+{% extends 'layout.jinja' %}
+{% block subtitle %}Login{% endblock %}
+
+{% block content %}
+
+<div class="prominent-form-wrapper">
+  <div class="prominent-form">
+    <a hx-boost="false" href="{{ url_for('login_redirect') }}">Log in via Keycloak</a>
+  </div>
+</div>
+{% endblock %}
