--- intersect_registry_service/app/ui/templates/pages/login-page.jinja
@@ -1,16 +1,0 @@
-{% extends 'layout.jinja' %}
-{% block subtitle %}Login{% endblock %}
-
-{% block content %}
-
-<div class="prominent-form-wrapper">
-  <form action="login" method="post" class="prominent-form" hx-disabled-elt="find button">
-    <span id="login-form-errors">
-      {% include 'login-page-error-partial.jinja' %}
-    </span>
-    <a hx-boost="false" href="/api/v1/auth/login">Log in via Keycloak</a>
-
-    <input type="hidden" name="csrf-token" value="{{ csrf_token }}" />
-  </form>
-</div>
-{% endblock %}
