+++ intersect_registry_service/app/ui/templates/pages/non-keycloak-login-page.jinja
@@ -1,0 +1,22 @@
+{% extends 'layout.jinja' %}
+{% block subtitle %}Login{% endblock %}
+
+{% block content %}
+
+<div class="prominent-form-wrapper">
+  <form action="login" method="post" class="prominent-form" hx-disabled-elt="find button" hx-boost="false">
+    <span id="login-form-errors">
+      {% include 'login-page-error-partial.jinja' %}
+    </span>
+    <label for="username"> Username</label>
+    <input type="text" id="login-username" name="username" placeholder="Enter your Username" value="{{username}}"
+      required />
+
+    <label for="password"> Password</label>
+    <input type="password" id="login-password" name="password" placeholder="Enter your Password" required />
+
+    <button type="submit">Submit</button>
+    <input type="hidden" name="csrf-token" value="{{ csrf_token }}" />
+  </form>
+</div>
+{% endblock %}
